research/requirements.txt
clickhouse-driver==0.2.9
matplotlib==3.10.0
psycopg2-binary==2.9.10
psutil==6.1.1
pytz==2024.2
six==1.17.0
tabulate==0.9.0
tzlocal==5.2
vertica-python==1.4.0
